Add resilience integration tests
Why:
We want to ensure kroxylicious can be restarted or crash, and have existing Kafka
clients recover in a timely fashion. Currently when it restarts it does not bind
ports for the brokers until a client connects to the bootstrap port and requests
metadata to drive port binding and build in-memory mappings of the upstream broker
addresses. Meaning clients can get stuck waiting for ports to be bound or SNI addresses
to be mapped.