feat(usernamespacefilter): UserNamespaceFilter (groupId and transactionalId only)
The role of the User Namespace Filter is to give the client a private space within the kafka cluster space that is isolated from other users sharing the cluster. Namespacing can be applied selectively to different resource types. This allows the possibility for some resource types (probably topics) to be shared between users while others (consumer group names and transactional ids) are private.

Signed-off-by: Keith Wall <kwall@apache.org>